Restructure identity hunter module
