+++ tests/test_utils.py
@@ -202,3 +202,5 @@ class TestConvertToJSON:
         assert circuit_contents[2]["target"] == 1
         assert circuit_contents[3]["phases"] == [0, 0]
         assert circuit_contents[3]["targets"] == [1, 0]
+
+
